Same final version. I just rename similarOffer.js to draw-offer-card.js
