+++ js/draw-offer-card.js
